mcp-server: rename the package and fold the stdio host into it
`@dxos/mcp-projection` named the transformation; the package is the DXOS MCP server —
the surface, its transports, and the gateway seam a host fills in — and it pairs with
`@dxos/mcp-client`. Renaming before the first publish keeps the npm name final.

The stdio layer moves out of the CLI: applying the response passes to the NDJSON line is
what makes a stdio host faithful, so it belongs with the surface, not with one caller.
