+++ .changeset/config.json
@@ -107,7 +107,7 @@
       "@dxos/log-store-idb",
       "@dxos/markdown",
       "@dxos/mcp-client",
-      "@dxos/mcp-projection",
+      "@dxos/mcp-server",
       "@dxos/merkle-search-tree",
       "@dxos/messaging",
       "@dxos/migrations",
