+++ packages/core/compute/mcp-server/moon.yml
