+++ packages/core/compute/mcp-server/package.json
@@ -1,8 +1,8 @@
 {
-  "name": "@dxos/mcp-projection",
+  "name": "@dxos/mcp-server",
   "version": "0.11.1",
   "private": true,
-  "description": "Projection of the operation registry onto an MCP tool and prompt surface.",
+  "description": "DXOS MCP server: the tool and prompt surface projected from the operation registry.",
   "homepage": "https://dxos.org",
   "bugs": "https://github.com/dxos/dxos/issues",
   "repository": {
