+++ packages/core/compute/mcp-server/src/Projection.test.ts
