+++ packages/core/compute/mcp-server/src/Projection.ts
