+++ packages/core/compute/mcp-server/src/Server.test.ts
