+++ packages/core/compute/mcp-server/src/Snapshot.ts
