+++ packages/core/compute/mcp-server/src/Wire.test.ts
