+++ packages/core/compute/mcp-server/src/Wire.ts
