+++ packages/core/compute/mcp-server/src/errors.ts
