+++ packages/core/compute/mcp-server/src/index.ts
@@ -7,5 +7,6 @@ export * as Projection from './Projection';
 export * as Server from './Server';
 export * as Snapshot from './Snapshot';
 export * as Space from './Space';
+export * as Stdio from './Stdio';
 export * as Wire from './Wire';
 export * from './errors';
