+++ packages/core/compute/mcp-server/tsconfig.json
