+++ packages/core/compute/mcp-server/vite.config.ts
