+++ tsconfig.all.json
@@ -301,7 +301,7 @@
       "path": "packages/core/compute/mcp-client/tsconfig.json"
     },
     {
-      "path": "packages/core/compute/mcp-projection/tsconfig.json"
+      "path": "packages/core/compute/mcp-server/tsconfig.json"
     },
     {
       "path": "packages/common/merkle-search-tree/tsconfig.json"
